moved everything that has to do with show/hide keyboard to CommonQuestionView and add a static showKeyboard method
